+++ lib/src/util/device_storage/device_storage_keys.dart
@@ -2,6 +2,7 @@ class DeviceStorageKeys {
   static const language = 'language';
   static const theme = 'theme';
   static const layoutHideNavLabels = 'layoutHideNavLabels';
+  static const layoutHideWallpaper = 'layoutHideWallpaper';
   static const logLevel = 'logLevel';
   static const logStack = 'logStack';
   static const initialAppStart = 'initialAppStart';