+++ lib/src/widgets/controls/card/glowing_border_container.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import '../../../util/theme_utils.dart';
+
 class GlowingBorderContainer extends StatelessWidget {
   // final double width;
   // final double height;
@@ -25,7 +27,12 @@ class GlowingBorderContainer extends StatelessWidget {
       // width: width,
       // height: height,
       margin: EdgeInsets.all(blurRadius),
-      decoration: createGlowingBoxDecoration(glowColor, themeData.scaffoldBackgroundColor, borderWidth: borderWidth, blurRadius: blurRadius),
+      decoration: createGlowingBoxDecoration(
+        glowColor,
+        ThemeUtils.isDarkMode(themeData) ? themeData.scaffoldBackgroundColor : themeData.cardTheme.color,
+        borderWidth: borderWidth,
+        blurRadius: blurRadius,
+      ),
       child: Center(child: child),
     );
   }